Fix bug in plane indexing for 2D3D moc sweeper

+++ src/cmdo/moc_sweeper_2d3d.cpp
@@ -33,7 +33,7 @@ namespace mocc {
             }
         }
 
-        flux_1g_ = flux_( blitz::Range::all(), (int)group );
+        flux_1g_ = flux_( blitz::Range::all(), group );
 
         // Perform inner iterations
         for( unsigned int inner=0; inner<n_inner_; inner++ ) {
